+++ W03StaticMesh_1/Week0v2/Engine/Source/Runtime/Engine/World.h
@@ -24,7 +24,7 @@ public:
     void CreateBaseObject();
     void ReleaseBaseObject();
     void RenderBaseObject();
-    void Tick(double DeltaTime);
+    void Tick(float DeltaTime);
     void Release();
     void Input();
     void SpawnObject(OBJECTS _Obj);
